Add direct VR connection for DX11, via HackerVR.
Uses same approach as prior code for 3Dmigoto, but altered to match the reshade environment.

Inserted into the runtime_d3d11.cpp for reshade at Present, it looks for the DoubleTex texture that SuperDepth3D creates, and if it exists, copies the contents into the shared surface.

Tested and running in VR, Katanga connects and shows the full SBS as stereo in VR.