Restructure dx11 path using c++ objects.
Not quite done here, but this is the basic structure to make this OO for the base object of 'vr', and then the subclass of vr_d3d11.

This keeps the common code like for the IPC in the base class, and the object specific ID3D11Texture2D in the subclass.

Doesn't work.  Needs to create the proper object in runtime_d3d11.