Rework CreateSharedTexture for doubleTex.
This brings the CreateTexture2D for shared surface here, as it will be same in all APIs.

Remove excess logging, leaving in FatalExit for things that should never fail.

Specifically convert the DoubleTex format via make_dxgi_format_normal, so that it's no longer Typeless.  If it's Typeless, then Katanga will fail when it tries to create the ShaderResourceView for display.  So by converting back to normal we create the shared surface as a normal DXGI_FORMAT_R8G8B8A8_UNORM.