Match backbuffer format when creating DoubleTex.
Looks for the DoubleTex texture being created, and if found, will change the format specified to match the backbuffer instead.
The SuperDepth is specifying only two types, RGBA8, or RGBA10, and neither will match all cases.  Rather than require specifying it per-game, this change just watches and creates a match.

It needs to match so that our copy out of backbuffer into DoubleTex via CopySubResourceRegion will succeed.

Tested as working in Yakuza0 which uses an unusual BGR8 format.  We correctly make DoubleTex as a Typeless format, to match how other parts of Reshade work.  3D works.