Add _doubletex reference to runtime object.
This will make it easier to determine when we are in stereo mode, without having to do the lookup all the time.
It gets initialized if the texture is created with the V__DoubleTex name, and destroyed when it is released.

By putting this in runtime, we can then use it in both runtime and DX11 specific runtime.