When we are in stereo mode- generate stereo screenshots.
We assume we are in stereo mode when the V__DoubleTex texture is available.
That fills in the _doubletex reference and we use that to switch to a double width stereo screenshot using _doubletex->impl instead of _backbuffer.
Eyes are also swapped when generating.

Tested in Rime and generates proper jps picture.