Move CaptureVRFrame to fix stuck eye bug.
Reshade will redraw the last seen backbuffer if MSAA is on, so that would overwrite whatever eye was drawn last and show up wrong in 3D Vision Direct.  Worked OK in VR mode.

This moves it to later in the pipeline, but is not much different.

However, this also means the GUI is no longer visible in 3D mode, because the frame with that data will no longer overwrite the backbuffer.