Draw twice more for the DoubleTex, during render_imgui.
We want to see the reshade Imgui window in both eyes, and this makes that happen by drawing the draw_data 3 times.  Once for the main backbuffer, and twice more for the DoubleTex, so that we retain the stereo, but get the UI over top.

Also required killing the scissorrect for the window, otherwise the right eye was clipped and not drawn.