Move CreateSharedTexture into vr_d3d11 to make it API specific.
The ending shared texture doesn't have to be DX11, so let's make this API specific in preparation for DX12 support.

This makes the call virtual, and is overriden in the API specific variants, to create and update their shared texture.  This makes the _shared_texture field IUnknown now instead, as the root class, and thus requires casting when used.